backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.4
python-dotenv==1.0.1
psycopg[binary]==3.2.1
pgvector==0.2.5
numpy==2.0.1
requests==2.32.3
python-multipart==0.0.9
